refactor: Swap to separate sqlite library, start updating server to use new db chore: Update to gradle 9

+++ app-model/src/commonMain/kotlin/me/dvyy/tasks/model/schema/JsonDAO.kt
@@ -3,8 +3,8 @@ package me.dvyy.tasks.model.schema
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.JsonElement
-import me.dvyy.syncengine.db.Transaction
-import me.dvyy.syncengine.db.WriteTransaction
+import me.dvyy.sqlite.Transaction
+import me.dvyy.sqlite.WriteTransaction
 import me.dvyy.syncengine.schema.JsonTable
 import org.intellij.lang.annotations.Language
 import kotlin.uuid.Uuid
@@ -30,29 +30,32 @@ class JsonDAO<T>(
     }
 
     context(tx: WriteTransaction)
+    fun create(
+        id: Uuid,
+        data: JsonElement,
+    ) {
+        tx.exec("INSERT INTO $table (id, data, owner) VALUES (?, jsonb(?), ?)", id, data.toString(), tx.identity)
+        tx.modified(table)
+    }
+
+    context(tx: WriteTransaction)
     fun patch(
         id: Uuid,
         data: T,
     ) = patch(id, json.encodeToString(serializer, data))
 
     context(tx: WriteTransaction)
-    fun create(
-        id: Uuid,
-        data: JsonElement,
-    ) {
-        tx.exec("INSERT INTO $table (id, data) VALUES (?, jsonb(?))", id, data.toString())
-        tx.modified(table)
-    }
-
-    context(tx: WriteTransaction)
     fun delete(id: Uuid) {
-        tx.exec("""DELETE FROM $table WHERE id = ?""", id)
+        tx.exec("DELETE FROM $table WHERE id = ? AND owner = ?", id, tx.identity)
         tx.modified(table)
     }
 
     context(tx: WriteTransaction)
     fun jsonSet(id: Uuid, path: String, value: String) {
-        tx.exec("UPDATE $table SET data = jsonb_set(data, ?, jsonb(?)) WHERE id = ?", path, value, id)
+        tx.exec(
+            "UPDATE $table SET data = jsonb_set(data, ?, jsonb(?)) WHERE id = ? AND owner = ?",
+            path, value, id, tx.identity
+        )
         tx.modified(table)
     }
 
@@ -68,9 +71,9 @@ class JsonDAO<T>(
             """
             UPDATE $table SET
             data = jsonb_patch(data, jsonb(?))
-            WHERE id = ?
+            WHERE id = ? AND owner = ?
             """.trimIndent(),
-            patchString, id
+            patchString, id, tx.identity
         )
         tx.modified(table)
     }
